Change script location
