--- Assets/Scripts/CodeGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 543be6ae6d6fd794bbd8e34fd9b2c179
