+++ Assets/Scripts/Debug.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: b2a0e47c7aa750c4f97adfb43fdcfd19
-TextScriptImporter:
+guid: 15ec7c3806f85714585e55f5d4608df5
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
